Dodat UserController
Dodat kontroler za entitet User

+++ java-backend/src/main/java/com/ftn/ac/rs/svtkvt2023/security/TokenUtils.java
@@ -76,7 +76,7 @@ public class TokenUtils {
                 .signWith(SignatureAlgorithm.HS512, secret).compact();
     }
 
-    public int getExpiredIn() {
-        return expiration.intValue();
+    public long getExpiredIn() {
+        return expiration;
     }
 }
